api/pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "snaplink-api"
version = "0.1.0"
description = "FastAPI signup API for Snaplink."
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.35,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "alembic>=1.16,<2.0",
  "pydantic-settings>=2.10,<3.0",
  "httpx>=0.28,<1.0",
  "boto3>=1.39,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
